Refactor schedule module
